Split the reducer into two parts
